requirements.txt
yt-dlp==2025.01.15
python-vlc==3.0.21203
spotipy==2.25.0
textual==1.0.0
rich==13.9.4

# Development dependencies
python-dotenv==1.0.1
pytest==8.3.4
httpx>=0.28.1
